Module Twostep

HOTP and TOTP algorithms for 2-step verification (for OCaml).

author Marco Aurélio da Silva

version 1.0.1

Pay attention that this library doesn't persist any HOTP counters neither tracks authenticated/verified valid OTP codes to protect against replay attacks. It's all up to you to build such defenses on top of this library with your underlying cloud-provided storage (NoSQL, relational databases, shared filesystems, etc).

module TOTP : sig ... end

Module for TOTP algorithm.

module HOTP : sig ... end

Module for HOTP algorithm.

module Internals : sig ... end

DON'T USE THAT MODULE, PRONE TO BREAKING CHANGES AND LIKELY UNSAFE/UNSECURE!